Fix: Ensure inputs are converted to float before passing to the network in Engine class

+++ hsi_setup.py
@@ -384,7 +384,7 @@ class Engine(object):
            
             #noisy_sigma = torch.zeros
             
-            outputs = self.net(inputs)
+            outputs = self.net(inputs.float())
            # outputs = torch.clamp(outputs, 0, 1)
             # loss = self.criterion(outputs, targets)
